mysite/requirements.txt
asgiref==3.7.2
autopep8==2.1.0
Django==5.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pycodestyle==2.11.1
pytesseract==0.3.10
sqlparse==0.4.4
typing_extensions==4.9.0
